Refactor to improve startup time
Use lazy loading technique to not require majority of the code and
the prettier package until needed. Also, don't use the babel pragma
as this seems to have a dramatic effect on load time. This means
using `module.exports` instead of `export const` and the like.
However, atom is still able to compile basic ES6 such as fat
arrow functions, `const`, and `let`, and this doesn't seem to
have much effect on load time.