feat($prettier-eslint): Integrate prettier-eslint
As discussed in:
- https://github.com/jlongster/prettier-atom/issues/57
- https://github.com/kentcdodds/prettier-eslint-atom/issues/28
- https://github.com/prettier/prettier/issues/758

This commit completely refactors the repo. We now allow using
prettier-eslint as an option so that you do not need to install
a separate plugin (prettier-eslint-atom).

Under the hood, almost all of the code has been refactored and
covered by Jest integration tests. We also implement Babel and
many of the devops/build configurations that Kent Dodds authored
in the prettier-eslint-atom repo.

Resolves #57