fix($formatOnSave): Handle null filePath
There may be cases where an editor will return a null filePath (such as when saving a new file).
Previously, doing so would cause an error because we assumed its presence when looking for the
nearest `.eslintignore` file. The solution is to short-circuit this check if the filePath is not
present, since there's no way it can be ignored if it doesn't have a path anyway.

Fixes #67