fix($config): Unset old config options causing errors
It seems that Atom may have a bug that is causing errors for some users. When we moved to a new
config schema, some users still have the old configuration options still set. This is preventing
them from loading prettier in some cases, because the options have incorrect values. It's difficult
to tell what is going on here and this fix may not work.

Fixes #72