build(package-scripts): Add `changelog` script
Add a `changelog` script for re-generating the CHANGELOG.md using the `conventional-changelog-cli`
package. For some reason, using the option of 1 release does nothing, so we use 0 and regenerate the
entire thing. If we want to manually amend CHANGELOG at any point, we need to fix this because it
will wipe those edits. We also don't inlcude the package in the dev dependencies due to a strange
Flow bug that won't allow it to parse a test file included with the package.