fix(formatOnSave): don't error when saving a new file
Fixes a bug where saving a brand new file would cause an error. This was because we were trying to
access the path of the editor's buffer's file when the file is actually null. Now we check for the
presence of the file before trying to access its path.

Fixes #79