feat(whitelist): force whitelist if provided
**What**: Update the logic of whitelist to be required if provided

**Why**: The use case of a whitelist is for people who have various
projects on their machine. Some of them they are able to use
`prettier-atom` on and others they are not. It's often easier to
maintain a list of projects they're allowed to use `prettier-atom` on
(whitelist) than it is to maintain a list of projects they are _not_
allowed to use `prettier-atom` on (blacklist). For these users, we say
that if you do provide a whitelist, then the files _must_ match the
whitelist or they will _not_ be formatted on save.

**How**:
- Add method `isWhitelistProvided` to `helpers`
- Add logic in `formatOnSaveIfAppropriate` for skipping if the whitelist
is provided and the file is not whitelisted
- Add note in config schema about this behavior