ci(travis): adjust caching option and implement .nvmrc
Travis now recommends just using "cache: yarn" instead of caching the node_modules directory. We
also add a .nvmrc file so that nvm and Travis will automatically detect which version of Node to
use. We can use such an edge version because we precompile the dist files so that they are very
backwards-compatible.