feat(eslint-fallbacks): fallback to user's prettier settings if cannot be inferred from eslint confi
Previously, if we couldn't infer a prettier setting from a user's eslint config, we fell back to
default prettier settings. Now, we take advantage of a new prettier-eslint feature that let's us
provide fallback options to use instead of the prettier defaults. This means that if you don't have
a semicolons rule in your eslint config, we will use the semicolon rule from your prettier-atom
settings.