refactor(src): change file organization
As the codebase has scaled, the src folder has become increasingly large and unwieldly.
Additionally, helpers and options had become quite extensive and lacked cohesion. This commit
refactors the various concerns into a nested folder structure. Functions that are only needed to
assist a specific domain are contained within that domain's folder. Helpers that interact with atom
are now in an atomInterface folder so that there is a clear interface to atom and so that it is easy
to mock accessing atom during testing. The same goes for helpers that interface with the Atom$Editor
object. Helpers are just those low-level functions that need to be shared between different domains.
We also move away from the strict "actual"/"expected" syntax for tests, although this may still be
used if expectation objects are considerably complex.