fix(formatOnSave): don't prevent user from saving if uncaught error
In unexpected situations where errors other than parsing errors during formatting are raised (the
latter already being handled), users are unable to save without disabling prettier because the Error
prevents the save from executing. Now we will wrap the logic inside of `formatOnSave` with a
try/catch that will catch this error and show a notification. This will inform the user that an
error occurred, but won't prevent saving their file.

Resolves #190