fix(handleError): properly handle alternative loc in Prettier errors
Prettier sometimes provides a different API for the `loc` property on the errors it throws. We were
only anticipating one type, where there is a `start` property nested in `loc`, but there's actually
another where the line and column are direct children of `loc` and there is no `start` property. We
now will properly handle either case.

Fixes #229