fix(display-of-errors): show syntax errors for unsaved files
Previously we were simply exiting early during the error handling process if there was no filepath
because the linter package requires a filepath in order to work. Now, we detect this early and
instead send the error message to a popup so that the user still gets some feedback.

See: https://github.com/steelbrain/linter/issues/1235

Fixes #235