fix(linter): resolve issue with linter not showing error messages properly
In some cases, the IndieLinter was `null` when it shouldn't have been. This was due to setting it
null when changing TextEditor objects, which was unnecessary and should not have been done. This
stops doing that so that there is always and IndieLinter available and thus we are always able to
properly set linter messages.

fix #318, #298