fix(set-text-via-diff): base use on option instead of flaky buffer range comparison
We were determining whether or not to use setTextViaDiff (more performant) versus the normal
setTextInBufferRange (less performant but allows replacing only a subset of text in a file if
desired) by whether or not the buffer range of the file equaled the buffer range of the text to be
transformed. However, in some cases such as when the editor automatically inserts an ending newline,
these two ranges can become unequal and we revert to using setTextInBufferRange. This is less
performant and not really necessary. The fix is to just pass an option when calling
`executePrettierOnBufferRange` from `formatOnSave`.