fix: properly use prettier config when formatting
We were relying on Prettier to look up its own settings, however, Prettier actually requires you
spoon-feeding it the options by calling its own `resolveConfig` method first and then feeding the
results of that as options to the `format` method. This bug was covered up because we were using
prettier-eslint which does its own lookup, but if you turn off the ESLint integration and rely on
Prettier directly, you get the bug.

fix #441