fix: ensure we always pass a relative path to `ignore` pkg
`atom.project.relativizePath` returns an absolute path when no Atom project can be determined (such
as when editing your Atom config file and many other cases). The `igore` package will then
hard-error because it is now very strict about getting only relative paths (they assume that the
filenames we are comparing the globs to should be relative to the file in which the globs live, but
we actually hold the globs in the prettier-atom options so this concept doesn't really apply so we
just arbitrarily choose to make relativize paths from the project's root dir). To fix, we relativize
to the parent dir of the file, which is somewhat arbitrary but should be fine.

Fixes #473