feat: Add option to enable format-on-save for files in node_modules
Currently it seems to be impossible to configure prettier-atom to 
format-on-save for files in node_modules, even by explicitly adding 
node_modules to the whitelist glob. This seems to be due to the fact 
that isFileFormattable defers to prettier's getFileInfo function, which 
by default will return `ignored: true` for files in node_modules, unless 
withNodeModules is passed in as an option. 

This PR introduces the ability to configure this behavior using a new 
ignoreNodeModules option. Suggestions on naming/copy for the option flag 
name/descriptions are very welcome, as are any suggestions people might 
have for testing/implementation.

I've defaulted the option to true for now to preserve existing behavior, 
though we may want to consider changing this default in a major version, 
because the current default can be a bit non-intuitive because users 
have plenty of other ways to explicitly exclude files from 
formatting-on-save (.prettierignore, exclusion glob, toggling 
format-on-save on/off, etc), so adding another layer of opaque default 
exclusions on top of that can make it seem like format-on-save is 
getting disabled arbitrarily on certain files (at least that was my 
impression at first until I looked deeper). FWIW the prettier extension 
for VSCode does format-on-save files in node_modules by default, and 
people don't seem to have any issues with it (presumably because those 
who care can simply configure it as a default exclusion).