build(node): use latest stable node version
Technically, the 10.x version is what Atom is using. However, sticking to this version for
development is now causing issues with updating to latest versions of development dependencies
(specifically, Jest) which hard errors because of an unsupported Node version. We just are going to
have to be careful that we don't introduce code that relies on newer node versions into what
actually gets compiled for use by Atom.